1.0 Seperated Overlay and Panel as requested
